requirements.txt
decorator==4.0.10
Django==1.9.8
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
virtualenv==15.0.1
wcwidth==0.1.7
restless==2.0.1
django-jwt-auth==0.0.2
django-nose==1.4.4
nose==1.3.7
ipdb==0.10.1
ipdbplugin==1.4.5
factory_boy==2.8.1